// apps/web/src/types/dashboard.types.ts
export type MonthlyComparisonTrend = "INCREASE" | "DECREASE" | "NO_CHANGE";

export interface DashboardMonthlyComparison {
  currentExpense: number;
  previousExpense: number;
  difference: number;
  percentage: number;
  trend: MonthlyComparisonTrend;
}